% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMO_plt_pca.R
\name{IMO_plt_pca}
\alias{IMO_plt_pca}
\title{PCA plot for massdataset.}
\usage{
IMO_plt_pca(
  obj,
  tag,
  center = T,
  scale = T,
  removeVar = 0.1,
  interactive = F,
  showImage = T
)
}
\arguments{
\item{obj}{mass_dataset; generated by tidymass.}

\item{tag}{character; MUST be one of the column names of sample_info, and case sensetive!}

\item{center}{logic; run ?PCAtools::pca in console.}

\item{scale}{logic; run ?PCAtools::pca in console.}

\item{removeVar}{float; run ?PCAtools::pca in console.}

\item{interactive}{logic; if true, generated a interactive 3D pca plot, if not, generated a 2d pca plot.}

\item{showImg}{logic; if true, print image on screen, if false, do not print.}
}
\value{
a list contains pca result and plots
}
\description{
PCA plot for massdataset.
}
\references{
https://github.com/kevinblighe/PCAtools
\dontrun{
IMO_plt_pca(obj = object, tag = group)
}
}
