% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMO_plt_rsd.R
\name{IMO_plt_rsd}
\alias{IMO_plt_rsd}
\title{rsd raw vs normalized.}
\usage{
IMO_plt_rsd(
  obj_old,
  obj_new,
  QC_tag,
  x_loc = c(120, 110),
  y_loc = c(15, 110),
  interactive = F,
  showImage = T
)
}
\arguments{
\item{obj_old}{A massdataset. raw data}

\item{obj_new}{A massdataset. normalized data}

\item{QC_tag}{character, QC tags in sample_info.}

\item{x_loc}{Integer Position of lables on the y-axis.}

\item{interactive}{logic; if true, generated a interactive 3D pca plot, if not, generated a 2d pca plot.}

\item{showImg}{logic; if true, print image on screen, if false, do not print.}
}
\value{
rsd_plt a 3D pca
}
\description{
rsd raw vs normalized.
}
