% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_serrf.R
\name{run_serrf}
\alias{run_serrf}
\title{SERRF functions.}
\usage{
run_serrf(obj, QC_tag, cluster_num = 8)
}
\arguments{
\item{obj}{mass_dataset, An object mass_dataset from massdataset package}

\item{QC_tag}{character, QC lable in sample_info column 'class'.}

\item{cluster_num}{how may cores do you want to use}
}
\value{
A SERRF normalized mass_dataset.
}
\description{
A QC based normalization method. Tidymass batch effect remove based on MS1 information of each batch, but SERRF didn't require MS1.
}
\references{
https://slfan2013.github.io/SERRF-online/
}
\author{
Shawn Wang
\email{shawnwang2016@126.com}
}
